Added Windows Cert PubSub sample and converted it to Windows Cert connection example
